requirements.txt
flask
lifxlan
boto3
bcrypt
PyJWT
botocore
python-dotenv
face_recognition
opencv-python
numpy